added dynamic framing and dynamic text flow to entity
Such that the blackboard/billboard background is no longer a graphic
but rather drawn directly on the canvas.  Similarly, the text can be
input in one big clump and the javascript will flow it into billboard.
Still makes some assumptions about layout, but that is easily fixed by
tweaking source.